dashboard finish

+++ src/components/Admin/Admin.jsx
@@ -4,6 +4,7 @@ import { useDispatch, useSelector } from 'react-redux'
 import { useEffect, useState } from 'react'
 import { getTrabajos, deleteTrabajo } from '../../redux/actions'
 import { Link } from 'react-router-dom'
+import Swal from 'sweetalert2'
 
 const Admin = () => {
 
@@ -15,10 +16,28 @@ const Admin = () => {
         dispatch(getTrabajos())
     }, [dispatch])
 
-    const handleDelete = async(id) => {
-        await dispatch(deleteTrabajo(id));
-        dispatch(getTrabajos())
-    }
+    const handleDelete = async (id) => {
+        Swal.fire({
+          title: "¿Deseas borrar este trabajo?",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          cancelButtonText: "Cancelar",
+          confirmButtonText: "Si, borrar"
+        }).then(async (result) => {
+          if (result.isConfirmed) {
+            await dispatch(deleteTrabajo(id));
+            dispatch(getTrabajos());
+      
+            Swal.fire({
+              title: "Borrado",
+              text: "El trabajo fue borrado exitosamente",
+              icon: "success"
+            });
+          }
+        });
+      };
     
     return(
         <div className={style.admin}>
@@ -48,15 +67,18 @@ const Admin = () => {
             {view === 'crear' ? (
                 <PostForm />
             ) : (
+                
+                <div className={style.cards}>
+                    {
                 trabajos.map((trabajo) => (
                     <div key={trabajo.id} className={style.card}>
+                        {trabajo.imagen && <img src={trabajo.imagen} alt={trabajo.nombre} />}
                         <h3>{trabajo.nombre}</h3>
                         <p>{trabajo.descripcion}</p>
-                        {trabajo.imagen && <img src={trabajo.imagen} alt={trabajo.nombre} />}
                         <button onClick={() => handleDelete(trabajo.id)}>Borrar</button>
                     </div>
                 ))
-            )}
+            }</div>)}
         </div>
         </div>
     );
